llm-cps-ac/src/chatbot/docker/requirements.txt
# Chatbot prerequisites (use dor Docker)

python-dotenv
openai
langchain
faiss-cpu
SQLAlchemy
tiktoken
streamlit
streamlit_chat
